client policy

+++ app/controllers/clients_controller.rb
@@ -1,14 +1,16 @@
 class ClientsController < ApplicationController
 
-  #include Pundit
-  #after_filter :verify_authorized
+  include Pundit
+  after_filter :verify_authorized
 
   def new
     @client = Client.new
+    authorize @client
   end
 
   def create
     @client = Client.new(client_params)
+    authorize @client
     if @client.save
       flash[:success] = 'Client created'
       redirect_to client_path(@client)
@@ -19,19 +21,22 @@ class ClientsController < ApplicationController
 
   def show
     @client = Client.find(params[:id])
+    authorize @client
   end
 
   def index
     @clients = Client.all
+    authorize @clients
   end
 
   def edit
     @client = Client.find(params[:id])
+    authorize @client
   end
 
   def update
     @client = Client.find(params[:id])
-#    authorize @client
+    authorize @client
     if @client.update_attributes(client_params)
       flash[:success] = 'Client updated'
       redirect_to client_path(@client)
@@ -42,7 +47,7 @@ class ClientsController < ApplicationController
 
   def destroy
     client = Client.find(params[:id])
-#    authorize client
+    authorize client
     client.destroy
     flash[:success] = 'Client destroyed.'
     redirect_to clients_url
